Windows 3rd-party library directories
Conditional logic has been added to the premake build file to automatically use
the correct libraries based on the platform and build configuration.

At the moment, an x86 build on windows is all that's really going to be supported,
but linux support will be in very soon. A few libraries are still not in the project
folder, so builds are still broken.