Logog headers included in project
The header files will be included so the project can be built without a bunch
of external downloads. This will continue to be standard practice as libraries
are added to the lib/ directory.

The premake file has been modified to only set the top-level "include" directory
as the search path. If any subfolders exists (such as with the logog lib), then
the code can still reference them as normal (i.e. #include "logog/logog.hpp").